Update Shelf_Edge_Clock_V1.ino
* changes

+++ Shelf_Edge_Clock_V1/Shelf_Edge_Clock_V1.ino
@@ -209,8 +209,9 @@ uint32_t generateNextColor(uint32_t inputColor, uint8_t shiftAmount, const char*
   uint8_t blue = inputColor & 0xFF;           // Extract the blue value
   
   // Calculate the next color by shifting the hue in a circular manner
-  float hue = rgbToHue(red, green, blue)   ;               // Convert RGB to HSL and extract the hue
-  hue = fmod(hue + shiftAmount, 360.0);                    // Shift the hue in a circular manner
+  float hue = rgbToHue(red, green, blue);                  // Convert RGB to HSL and extract the hue
+  //hue = fmod(hue + shiftAmount, 360.0);                    // Shift the hue in a circular manner
+  hue = fmod(hue + static_cast<float>(shiftAmount), 360.0f);
   uint32_t nextColor = hueToRgb(hue, red, green, blue);    // Convert back to RGB
  
   Serial.print("current HUE: ");  Serial.println(hue);
@@ -263,33 +264,33 @@ uint32_t hueToRgb(float hue, uint8_t& red, uint8_t& green, uint8_t& blue) {
   switch (sector) {
     case 0:
       red = 255;
-      green = 255 * fraction;
+      green = static_cast<uint8_t>(255 * fraction);
       blue = 0;
       break;
     case 1:
-      red = 255 * q;
+      red = static_cast<uint8_t>(255 * q);
       green = 255;
       blue = 0;
       break;
     case 2:
       red = 0;
       green = 255;
-      blue = 255 * fraction;
+      blue = static_cast<uint8_t>(255 * fraction);
       break;
     case 3:
       red = 0;
-      green = 255 * q;
+      green = static_cast<uint8_t>(255 * q);
       blue = 255;
       break;
     case 4:
-      red = 255 * fraction;
+      red = static_cast<uint8_t>(255 * fraction);
       green = 0;
       blue = 255;
       break;
     default:
       red = 255;
       green = 0;
-      blue = 255 * q;
+      blue = static_cast<uint8_t>(255 * q);
       break;
   }
